LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGGeometryElement.getTotalLength-02-expected.txt


FAIL SVGGeometryElement.getTotalLength: 'display:none' assert_equals: expected 40 but got 0
